+++ src/StorageSync/StorageSync/CloudEndpoint/NewCloudEndpointCommand.cs
@@ -159,14 +159,14 @@ namespace Microsoft.Azure.Commands.StorageSync.CloudEndpoint
         public string StorageAccountTenantId { get; set; }
 
         /// <summary>
-        /// Gets or sets the change enumeration interval days.
+        /// Gets or sets the change enumeration interval day.
         /// </summary>
-        /// <value>The change enumeration interval days.</value>
+        /// <value>The change enumeration interval day.</value>
         [Parameter(Mandatory = false,
                    ValueFromPipelineByPropertyName = false,
-                   HelpMessage = HelpMessages.ChangeEnumerationIntervalDaysParameter)]
+                   HelpMessage = HelpMessages.ChangeEnumerationIntervalDayParameter)]
         [ValidateRange(1, 20)]
-        public int? ChangeEnumerationIntervalDays { get; set; }
+        public int? ChangeEnumerationIntervalDay { get; set; }
 
         /// <summary>
         /// Gets or sets as job.
@@ -247,7 +247,7 @@ namespace Microsoft.Azure.Commands.StorageSync.CloudEndpoint
                     StorageAccountResourceId = StorageAccountResourceId,
                     AzureFileShareName = AzureFileShareName,
                     StorageAccountTenantId = (StorageAccountTenantId ?? AzureContext.Tenant.Id),
-                    ChangeEnumerationIntervalDays = ChangeEnumerationIntervalDays
+                    ChangeEnumerationIntervalDays = ChangeEnumerationIntervalDay
                 };
 
                 string resourceGroupName = ResourceGroupName ?? ParentObject?.ResourceGroupName ?? parentResourceIdentifier.ResourceGroupName;